moves display to SDK
